Move make_model_image
